==> pipeline/3_join_parent.ipynb <==
# Feature engineering: features from parent posts

## Read raw questions & answers data

`data/raw/stackoverflow_amazon_q.csv` is actually a tsv file.

import pandas as pd

RAW_Q = r'../data/raw/stackoverflow_amazon_q.csv'
RAW_A = r'../data/raw/stackoverflow_amazon_a.csv'
PIPE_C_LINGU = r'../data\pipeline\c_heuristic_features.csv'
PIPE_C_Q_TAGS = r'../data/pipeline/cq_parent_feats.csv'
PIPE_C_A_TAGS = r'../data/pipeline/ca_parent_feats.csv'
PIPE_A_TAGS = r'../data/pipeline/a_parent_feats.csv'

df_raw_questions = pd.read_table(RAW_Q, delimiter='\t', error_bad_lines=False, warn_bad_lines=False)

df_raw_questions.columns

df_raw_questions.info()

df_raw_questions

df_raw_questions['Tags']

df_raw_answers = pd.read_csv(RAW_A, error_bad_lines=False, warn_bad_lines=False)

df_raw_answers.columns

df_raw_answers.info()

## Read outdated comments & answers

Data for answers is WIP.

It seems I got the opposite value for 'of_answer' column: a_id is in questions IDs, q_id is in answers IDs.

df_outdated_comments = pd.read_csv(PIPE_C_LINGU)

df_outdated_comments.columns

a_id_set = set(df_outdated_comments[df_outdated_comments['of_answer'] == False]['PostId'])
q_id_set = set(df_outdated_comments[df_outdated_comments['of_answer'] == True]['PostId'])

sum(df_raw_answers['Id'].isin(a_id_set))

df_raw_questions.columns

df_raw_answers.columns

df_outdated_comments.columns

sum(df_raw_questions['Id'].isin(q_id_set))

df_used_ans = df_raw_answers[df_raw_answers['Id'].isin(q_id_set)]
a_pid_set = set(df_used_ans['ParentId'])
df_used_que = df_raw_questions[df_raw_questions['Id'].isin(a_id_set.union(a_pid_set))]

## Join tables

First, join the answers with features in questions. (only the tags & FavoriteCount & ClosedDate)

Then join the comments with the question table and joined answer table.

Q_FEAT_COLS = ['Id', 'CreationDate', 'Score', 'ViewCount',
               'LastEditDate', 'LastActivityDate', 'Title', 'Tags', 'AnswerCount', 'CommentCount',
               'ClosedDate', 'FavoriteCount']
A_FEAT_COLS = ['Id', 'ParentId', 'CreationDate', 'Score', 'Body',
               'LastActivityDate', 'CommentCount']

df_used_ans = df_used_ans[A_FEAT_COLS]
df_used_que = df_used_que[Q_FEAT_COLS]

df_comments_a = df_outdated_comments[df_outdated_comments['of_answer'] == True]
df_comments_q = df_outdated_comments[df_outdated_comments['of_answer'] == False]

df_joined_ans.info()

df_joined_ans = pd.merge(df_used_ans, df_used_que, how='left', 
                         left_on='ParentId', right_on='Id',
                        suffixes=('_ans','_que'))

df_comments_q.columns

df_used_que.columns

None of question IDs is a PostId in `df_comments_q`. (fixed; now it's right.) 

sum(df_used_que.Id.isin(set(df_comments_q.PostId)))

sum(df_used_que.Id.isin(set(df_comments_a.PostId)))

df_joined_com_q = pd.merge(df_comments_q, df_used_que, how='left',
                          left_on='PostId', right_on='Id', suffixes=('', '_que'))
df_joined_com_q.columns

df_joined_com_a = pd.merge(df_comments_a, df_joined_ans, how='left',
                          left_on='PostId', right_on='Id_ans')
df_joined_com_a.columns

The same procedure for answers.

df_ans_ling = pd.read_csv('../data/pipeline/a_heuristic_feats.csv')
df_ans_ling = df_ans_ling[['Id', 'PostTypeId', 'ParentId', 'CreationDate', 'Score',
       'Body', 'OwnerUserId', 'OwnerDisplayName', 'LastActivityDate',
       'CommentCount', 'ContentLicense', 'cnt_keywords', 'subject',
       'punctuation', 'negative_statement', 'sentence', 'subj_irrel',
       'include_irrel']]

df_joined_ans_ling = pd.merge(df_ans_ling, df_raw_questions[Q_FEAT_COLS], how='left', 
                         left_on='ParentId', right_on='Id',
                        suffixes=('','_que'))

df_joined_ans_ling.info()

## Save the joined tables

df_joined_com_a.to_csv(PIPE_C_A_TAGS, index=False)
df_joined_com_q.to_csv(PIPE_C_Q_TAGS, index=False)
df_joined_ans_ling.to_csv(PIPE_A_TAGS, index=False)
