==> pipeline/5_Merge_and_scale_manual_check.ipynb <==
import pandas as pd
import numpy as np
from collections import Counter
import re

# read comments cvs
# df_comments = pd.read_csv("Manual checking - comments.csv")
# df_comments = pd.read_csv("./data/pipeline/cq_parent_feats.csv")
df_comments = pd.read_csv("./data/pipeline/cq_parent_feats.csv")
df_comments

df_comments['from_dataset']='comments_que'

df_comments.info()

df_comments['Id_que']

# df_answers = pd.read_csv("./data/pipeline/ca_parent_feats.csv")
df_answers = pd.read_csv("./data/pipeline/ca_parent_feats.csv")
# df_answers.columns
df_answers = df_answers.drop('Body',axis = 1)
# df_answers['Id_ans']==df_answers['Id']

df_comments['from_dataset']='comments_ans'

df_answers.columns

df_answers_a = pd.read_csv("./data/pipeline/a_parent_feats.csv")
df_answers_a = df_answers_a.drop('Body',axis = 1)

df_answers_a['from_dataset']='answers'

df_answers_a.columns

df_answers_a.info()

df_comments.columns

set(df_comments.columns)-set(df_answers.columns)

set(df_answers.columns)-set(df_comments.columns)

print(len(df_comments.columns))
print(len(df_answers.columns))

df = pd.concat([df_comments,df_answers],axis = 0,ignore_index = True)

df = pd.concat([df,df_answers_a],axis = 0,ignore_index = True)

df.info()

#CommunityOwnedDate      0 non-null float64
#drop the columns completely lost
df = df.drop('CommunityOwnedDate',axis = 1)

def Find(string):
    if string!=string:
        return 0
    regex = r"(?i)\b((?:https?://|www\d{0,3}[.]|[a-z0-9.\-]+[.][a-z]{2,4}/)(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()\[\]{};:'\".,<>?«»“”‘’]))"
    url = re.findall(regex,string)
    url_list=[x[0] for x in url]
    if len(url_list) != 0:
        return 1
    else:
        return 0
df['share_link']= df['Text'].apply(lambda x:Find(x))

def get_tags(tags):
#     print(tags)
    if tags != tags:
        return None
    else:
        tags = tags[1:-1]
        tags = tags.split('><')
        res = []
        for item in tags:
            if ('aws' in item) or ('amazon' in item):
                if item != 'aws' and item !='amazon-web-services':
                    res.append(item)
        if res == []:
            return None
        return res
df['aws_related_tags']= df['Tags'].apply(lambda x:get_tags(x))

def count_tag_num(tags):
    if tags == None:
        return None
    return len(tags)
df['tag_num']= df['aws_related_tags'].apply(lambda x:count_tag_num(x))

pd.unique(df['tag_num'])

def get_tags(tags):
    if tags != tags:
        return None
    else:
        tags = tags[1:-1]
        tags = tags.split('><')
        res = []
        for item in tags:
            if ('aws' in item) or ('amazon' in item):
                if item != 'aws' and item !='amazon-web-services':
                    res.append(item)
        if res == []:
            return None
        return res
df['aws_related_tags']= df['Tags'].apply(lambda x:get_tags(x))

def get_code(text):
    if text!=text:
        return 0
    code = re.findall('`([^"]*)`',text)
    if code ==[]:
        return 0
    else:
        return 1
df['share_code']= df['Text'].apply(lambda x:get_code(x))

df

# read answers cvs
# df_answers = pd.read_csv("Manual checking - answers.csv")
df_answers_checked = pd.read_csv("./data/pipeline/Manual checking - answers (1).csv")
df_answers_checked = df_answers_checked[:100]

df_answers_checked = df_answers_checked[['answer_id','outdated (manually checked)']]

df = df.merge(df_answers_checked,left_on='Id_ans',right_on='answer_id',how = 'left')

# read comments cvs
# df_comments = pd.read_csv("Manual checking - comments.csv")
df_comments_checked = pd.read_csv("./data/pipeline/Manual checking - comments(new).csv")
df_comments_checked = df_comments_checked[:194]


# df_comments_checked 

df_comments_checked = df_comments_checked[['Id','outdated (manually checked)']]

df = df.merge(df_comments_checked,left_on='Id',right_on='Id',how = 'left')

def get_outdated(x,y):
    if x == x:
        return x
    elif y == y:
        return y
    else:
        return None
df['outdated (manually checked)']= df.apply(lambda x:get_outdated(x['outdated (manually checked)_x'],x['outdated (manually checked)_y']),axis =1)

def train_or_test(manual):
    if manual == None:
        return 'Test'
    else:
        return 'Train'
df['is_training']= df['outdated (manually checked)'].apply(lambda x:train_or_test(x))
 

df = df.drop(['outdated (manually checked)_x','outdated (manually checked)_y','answer_id'],axis = 1)

df = df.drop(['tag_num'],axis =1)

def get_keywords(text):
    if text != text:
        return None
    text_dict = ['outdated','deprecated','obsolete','out of date','not deprecated','old version','probably outdated','remove','instead']
    res = []
    for item in text_dict:
        if item in text:
            res.append(item)
    res = set(res)
    if 'not deprecated' in res and 'deprecated' in res:
        res.remove('deprecated')
    return res
df['keywords']= df['Text'].apply(lambda x:get_keywords(x))

df.columns

df.to_csv("./data/pipeline/1550_rows_45_columns.csv")

df
